chore(rpc): split cache
